Split admin operations surfaces
