+++ frontend/components/admin/JobAuditTable.tsx
@@ -1,177 +1,15 @@
 'use client';
 
-import Link from 'next/link';
-import { type ReactNode, Fragment, useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import clsx from 'clsx';
-import type { AdminJobAuditRecord, AdminJobOutcome } from '@/server/admin-job-audit';
+import type { AdminJobAuditRecord } from '@/server/admin-job-audit';
+import { JobAuditRows } from '@/components/admin/job-audit/JobAuditRows';
 import { Button } from '@/components/ui/Button';
 
 interface JobAuditTableProps {
   initialJobs: AdminJobAuditRecord[];
   initialCursor: string | null;
   filtersQuery: string;
-}
-
-const dateTimeFormatter = new Intl.DateTimeFormat('en-GB', {
-  day: '2-digit',
-  month: '2-digit',
-  year: 'numeric',
-  hour: '2-digit',
-  minute: '2-digit',
-  second: '2-digit',
-  hourCycle: 'h23',
-  timeZone: 'Europe/Paris',
-});
-
-function formatCurrency(amountCents: number, currency: string) {
-  try {
-    return new Intl.NumberFormat('en-US', {
-      style: 'currency',
-      currency: currency.toUpperCase(),
-    }).format(amountCents / 100);
-  } catch {
-    return `${(amountCents / 100).toFixed(2)} ${currency.toUpperCase()}`;
-  }
-}
-
-function formatDate(value: string | null | undefined) {
-  if (!value) return '—';
-  try {
-    const date = new Date(value);
-    if (Number.isNaN(date.getTime())) return value;
-    const parts = Object.fromEntries(
-      dateTimeFormatter
-        .formatToParts(date)
-        .filter((part) => part.type !== 'literal')
-        .map((part) => [part.type, part.value])
-    );
-    return `${parts.day}/${parts.month}/${parts.year} ${parts.hour}:${parts.minute}:${parts.second}`;
-  } catch {
-    return value;
-  }
-}
-
-function outcomeMeta(outcome: AdminJobOutcome) {
-  const base = 'inline-flex items-center rounded-full border px-2 py-0.5 text-xs font-semibold uppercase tracking-micro';
-  if (outcome === 'failed_action_required') {
-    return {
-      label: 'Action required',
-      className: clsx(base, 'border-error-border bg-error-bg text-error'),
-    };
-  }
-  if (outcome === 'refunded_failure_resolved') {
-    return {
-      label: 'Refunded failure',
-      className: clsx(base, 'border-info-border bg-info-bg text-info'),
-    };
-  }
-  if (outcome === 'completed') {
-    return {
-      label: 'Completed',
-      className: clsx(base, 'border-success-border bg-success-bg text-success'),
-    };
-  }
-  if (outcome === 'in_progress') {
-    return {
-      label: 'In progress',
-      className: clsx(base, 'border-warning-border bg-warning-bg text-warning'),
-    };
-  }
-  return {
-    label: 'Unknown',
-    className: clsx(base, 'border-border bg-muted text-text-secondary'),
-  };
-}
-
-function technicalStatusBadge(status?: string | null) {
-  if (!status) return null;
-  const normalized = status.toLowerCase();
-  const base = 'inline-flex items-center rounded-full border px-2 py-0.5 text-xs font-semibold uppercase tracking-micro';
-  if (normalized === 'completed') {
-    return <span className={clsx(base, 'border-success-border bg-success-bg text-success')}>completed</span>;
-  }
-  if (normalized === 'failed') {
-    return <span className={clsx(base, 'border-error-border bg-error-bg text-error')}>failed</span>;
-  }
-  if (normalized === 'running' || normalized === 'queued' || normalized === 'pending') {
-    return <span className={clsx(base, 'border-warning-border bg-warning-bg text-warning')}>{normalized}</span>;
-  }
-  return <span className={clsx(base, 'border-info-border bg-info-bg text-info')}>{normalized}</span>;
-}
-
-function displayMeta(job: AdminJobAuditRecord) {
-  const surface = job.surface ?? 'video';
-  if (surface === 'image') {
-    return {
-      readyLabel: 'Image ready',
-      placeholderLabel: 'Placeholder image',
-      missingLabel: 'Missing image',
-      linkLabel: 'Open image',
-    };
-  }
-  if (surface === 'character') {
-    return {
-      readyLabel: 'Character ready',
-      placeholderLabel: 'Placeholder image',
-      missingLabel: 'Missing image',
-      linkLabel: 'Open character',
-    };
-  }
-  if (surface === 'angle') {
-    return {
-      readyLabel: 'Angle ready',
-      placeholderLabel: 'Placeholder image',
-      missingLabel: 'Missing image',
-      linkLabel: 'Open angle',
-    };
-  }
-  return {
-    readyLabel: 'Video ready',
-    placeholderLabel: 'Placeholder asset',
-    missingLabel: 'Missing video',
-    linkLabel: 'Open video',
-  };
-}
-
-function IntegrityPill({
-  ok,
-  label,
-  helper,
-  warningLabel,
-}: {
-  ok: boolean;
-  label: string;
-  helper: string;
-  warningLabel: string;
-}) {
-  return (
-    <div className="space-y-1">
-      <span
-        className={clsx(
-          'inline-flex items-center rounded-full border px-2 py-0.5 text-xs font-semibold uppercase tracking-micro',
-          ok ? 'border-success-border bg-success-bg text-success' : 'border-warning-border bg-warning-bg text-warning'
-        )}
-      >
-        {ok ? label : warningLabel}
-      </span>
-      <p className="text-xs text-text-secondary">{helper}</p>
-    </div>
-  );
-}
-
-function DetailCard({
-  title,
-  children,
-}: {
-  title: string;
-  children: ReactNode;
-}) {
-  return (
-    <div className="rounded-2xl border border-hairline bg-bg/70 p-4 text-xs text-text-secondary">
-      <p className="mb-2 text-[11px] font-semibold uppercase tracking-[0.18em] text-text-muted">{title}</p>
-      <div className="space-y-1.5">{children}</div>
-    </div>
-  );
 }
 
 export function AdminJobAuditTable({ initialJobs, initialCursor, filtersQuery }: JobAuditTableProps) {
@@ -441,245 +279,15 @@ export function AdminJobAuditTable({ initialJobs, initialCursor, filtersQuery }:
               </tr>
             </thead>
             <tbody>
-              {visibleJobs.length === 0 ? (
-                <tr>
-                  <td colSpan={6} className="px-4 py-10 text-center text-text-secondary">
-                    No jobs found for the current filters.
-                  </td>
-                </tr>
-              ) : (
-                visibleJobs.map((job) => {
-                  const meta = outcomeMeta(job.outcome);
-                  const display = displayMeta(job);
-                  const isExpanded = expandedJobId === job.jobId;
-                  const refundEligible =
-                    job.paymentStatus?.includes('paid_wallet') && job.totalChargeCents > 0 && job.totalRefundCents === 0;
-
-                  return (
-                    <Fragment key={job.jobId}>
-                      <tr className="border-t border-hairline align-top transition hover:bg-bg">
-                        <td className="px-4 py-3">
-                          <div className="space-y-2">
-                            <div>
-                              <p className="text-sm font-medium text-text-primary">{formatDate(job.createdAt)}</p>
-                              <p className="text-xs text-text-secondary">Updated {formatDate(job.updatedAt)}</p>
-                            </div>
-                            {job.archived ? (
-                              <span className="inline-flex items-center rounded-full border border-warning-border bg-warning-bg px-2 py-0.5 text-xs font-semibold uppercase tracking-micro text-warning">
-                                Archived candidate
-                              </span>
-                            ) : null}
-                          </div>
-                        </td>
-
-                        <td className="px-4 py-3">
-                          <div className="min-w-0 space-y-2">
-                            <p className="font-mono text-xs text-text-primary break-all">{job.jobId}</p>
-                            {job.providerJobId ? (
-                              <p className="font-mono text-[11px] text-text-muted break-all">{job.providerJobId}</p>
-                            ) : null}
-                            {job.failureReason ? (
-                              <p className="line-clamp-2 text-xs text-text-secondary">{job.failureReason}</p>
-                            ) : job.message ? (
-                              <p className="line-clamp-2 text-xs text-text-secondary">{job.message}</p>
-                            ) : null}
-                            {job.outputUrl ? (
-                              <a
-                                href={job.outputUrl}
-                                target="_blank"
-                                rel="noreferrer"
-                                className="inline-flex text-xs font-medium text-brand underline-offset-2 hover:underline"
-                              >
-                                {display.linkLabel}
-                              </a>
-                            ) : null}
-                          </div>
-                        </td>
-
-                        <td className="px-4 py-3">
-                          <div className="space-y-2 text-xs">
-                            {job.userId ? (
-                              <Link href={`/admin/users/${job.userId}`} className="font-mono text-text-primary underline-offset-2 hover:underline">
-                                {job.userId}
-                              </Link>
-                            ) : (
-                              <p className="text-text-secondary">No user attached</p>
-                            )}
-                            <div>
-                              <p className="font-medium text-text-primary">{job.engineLabel ?? 'Unknown engine'}</p>
-                              <p className="text-text-secondary">
-                                Surface {job.surface ?? 'video'}
-                                {job.durationSec != null ? ` · ${job.durationSec}s` : ''}
-                              </p>
-                            </div>
-                          </div>
-                        </td>
-
-                        <td className="px-4 py-3">
-                          <div className="space-y-2">
-                            <div className="flex flex-wrap gap-2">
-                              <span className={meta.className}>{meta.label}</span>
-                              {technicalStatusBadge(job.status)}
-                            </div>
-                            <p className="text-xs text-text-secondary">Progress {job.progress ?? 0}%</p>
-                            {job.failureOrigin ? (
-                              <p className="text-xs text-text-secondary">Origin {job.failureOrigin}</p>
-                            ) : null}
-                          </div>
-                        </td>
-
-                        <td className="px-4 py-3">
-                          <div className="space-y-3">
-                            <IntegrityPill
-                              ok={job.paymentOk}
-                              label="Payment OK"
-                              warningLabel="Payment drift"
-                              helper={`Net ${formatCurrency(job.netChargeCents, job.currency)} · expected ${formatCurrency(job.finalPriceCents, job.currency)}`}
-                            />
-                            <IntegrityPill
-                              ok={job.falOk}
-                              label="Fal OK"
-                              warningLabel="Fal issue"
-                              helper={`Latest ${job.falStatus ?? '—'} · logs ${job.falLogCount}`}
-                            />
-                            <IntegrityPill
-                              ok={job.hasOutput}
-                              label={display.readyLabel}
-                              warningLabel={job.isPlaceholderOutput ? display.placeholderLabel : display.missingLabel}
-                              helper={job.outputUrl ? 'Output linked' : 'No output URL recorded'}
-                            />
-                          </div>
-                        </td>
-
-                        <td className="px-4 py-3">
-                          <div className="flex flex-col items-end gap-2">
-                            <Button
-                              type="button"
-                              size="sm"
-                              variant="outline"
-                              onClick={() => setExpandedJobId((prev) => (prev === job.jobId ? null : job.jobId))}
-                              className="border-border bg-surface"
-                            >
-                              {isExpanded ? 'Hide details' : 'Details'}
-                            </Button>
-
-                            {job.providerJobId && (!job.hasOutput || job.status !== 'completed') ? (
-                              <Button
-                                type="button"
-                                size="sm"
-                                variant="outline"
-                                onClick={() => void handleResync(job.jobId)}
-                                disabled={syncingJobId === job.jobId}
-                                className="border-border bg-surface"
-                              >
-                                {syncingJobId === job.jobId ? 'Syncing…' : 'Resync Fal'}
-                              </Button>
-                            ) : null}
-
-                            {refundEligible ? (
-                              <Button
-                                type="button"
-                                size="sm"
-                                variant="outline"
-                                onClick={() => handleRefund(job.jobId)}
-                                className="border-error-border bg-error-bg text-error hover:text-error"
-                              >
-                                Refund tokens
-                              </Button>
-                            ) : null}
-
-                            {job.archived ? (
-                              <Button
-                                type="button"
-                                size="sm"
-                                variant="outline"
-                                onClick={() => handleRestore(job.jobId)}
-                                className="border-border bg-surface"
-                              >
-                                Bring online
-                              </Button>
-                            ) : null}
-                          </div>
-                        </td>
-                      </tr>
-
-                      {isExpanded ? (
-                        <tr className="border-t border-hairline bg-surface/50">
-                          <td colSpan={6} className="px-4 py-4">
-                            <div className="grid gap-4 xl:grid-cols-4">
-                              <DetailCard title="Output & Routing">
-                                <p>Surface: {job.surface ?? 'video'}</p>
-                                <p>Engine: {job.engineLabel ?? 'Unknown engine'}</p>
-                                <p>Provider ID: {job.providerJobId ?? '—'}</p>
-                                <p>Hero render: {job.heroRenderId ?? '—'}</p>
-                                <p>Thumb: {job.thumbUrl ?? '—'}</p>
-                                <p>Render count: {job.renderCount}</p>
-                                {job.outputUrl ? (
-                                  <a href={job.outputUrl} target="_blank" rel="noreferrer" className="font-medium text-brand underline-offset-2 hover:underline">
-                                    {display.linkLabel}
-                                  </a>
-                                ) : null}
-                              </DetailCard>
-
-                              <DetailCard title="Billing">
-                                <p>Status: {job.paymentStatus ?? '—'}</p>
-                                <p>Expected: {formatCurrency(job.finalPriceCents, job.currency)}</p>
-                                <p>Charges: {formatCurrency(job.totalChargeCents, job.currency)}</p>
-                                <p>Refunds: {formatCurrency(job.totalRefundCents, job.currency)}</p>
-                                <p>Net: {formatCurrency(job.netChargeCents, job.currency)}</p>
-                                <p>Charge count: {job.chargeCount}</p>
-                                <p>Refund count: {job.refundCount}</p>
-                                {job.receipts.length ? (
-                                  <div className="pt-1">
-                                    <p className="font-semibold text-text-primary">Recent receipts</p>
-                                    <ul className="mt-1 space-y-1">
-                                      {job.receipts.slice(0, 4).map((receipt) => (
-                                        <li key={receipt.id} className="font-mono text-[11px]">
-                                          {receipt.type} · {formatCurrency(receipt.amountCents, receipt.currency)} · {formatDate(receipt.createdAt)}
-                                        </li>
-                                      ))}
-                                    </ul>
-                                  </div>
-                                ) : null}
-                              </DetailCard>
-
-                              <DetailCard title="Fal Diagnostics">
-                                <p>Latest status: {job.falStatus ?? '—'}</p>
-                                <p>Updated: {formatDate(job.falUpdatedAt)}</p>
-                                <p>Failure at: {formatDate(job.failureAt)}</p>
-                                <p>Failure origin: {job.failureOrigin ?? '—'}</p>
-                                <p>Reason: {job.failureReason ?? '—'}</p>
-                                <p>Refunded: {job.isRefunded ? 'Yes' : 'No'}</p>
-                                <p>Refund note: {job.refundReason ?? '—'}</p>
-                                <p>Fal log entries: {job.falLogCount}</p>
-                              </DetailCard>
-
-                              <DetailCard title="Timeline">
-                                {job.timeline.length ? (
-                                  <ul className="space-y-2">
-                                    {job.timeline.map((event, index) => (
-                                      <li key={`${job.jobId}-${event.at}-${event.kind}-${index}`}>
-                                        <p className="font-mono text-[11px] text-text-primary">{formatDate(event.at)}</p>
-                                        <p>
-                                          {event.source} · {event.kind}
-                                        </p>
-                                        <p>{event.summary}</p>
-                                        {event.details ? <p>{event.details}</p> : null}
-                                      </li>
-                                    ))}
-                                  </ul>
-                                ) : (
-                                  <p>No timeline events.</p>
-                                )}
-                              </DetailCard>
-                            </div>
-                          </td>
-                        </tr>
-                      ) : null}
-                    </Fragment>
-                  );
-                })
-              )}
+              <JobAuditRows
+                expandedJobId={expandedJobId}
+                jobs={visibleJobs}
+                onRefund={handleRefund}
+                onRestore={handleRestore}
+                onResync={(jobId) => void handleResync(jobId)}
+                onToggleExpanded={(jobId) => setExpandedJobId((prev) => (prev === jobId ? null : jobId))}
+                syncingJobId={syncingJobId}
+              />
             </tbody>
           </table>
         </div>
