+++ frontend/components/admin/PlaylistsManager.tsx
@@ -1,35 +1,29 @@
 "use client";
 
-import Image from 'next/image';
 import {
-  Fragment,
   useCallback,
   useEffect,
   useMemo,
-  useRef,
   useState,
   useTransition,
-  type DragEvent,
   type FormEvent,
 } from 'react';
 import clsx from 'clsx';
-import { Button } from '@/components/ui/Button';
 import { authFetch } from '@/lib/authFetch';
-import { PlaylistCreateForm } from '@/components/admin/playlists/PlaylistCreateForm';
-import { PlaylistDetailsPanel } from '@/components/admin/playlists/PlaylistDetailsPanel';
 import { PlaylistFeedbackBanners } from '@/components/admin/playlists/PlaylistFeedbackBanners';
+import { PlaylistOrderDirtyBar } from '@/components/admin/playlists/PlaylistItemsSection';
+import { PlaylistsManagerToolbar } from '@/components/admin/playlists/PlaylistsManagerToolbar';
+import { PlaylistsManagerSelectionPanel } from '@/components/admin/playlists/PlaylistsManagerSelectionPanel';
 import { PlaylistsSidebar } from '@/components/admin/playlists/PlaylistsSidebar';
-import { StatusPill } from '@/components/admin/playlists/PlaylistStatusPill';
+import { usePlaylistDragReorder } from '@/components/admin/playlists/usePlaylistDragReorder';
 import {
   buildFamilyHelpers,
   buildPlaylistUpdateFromItems,
-  getPlaceholderThumb,
   getPlaylistGroup,
-  slugify,
   sortItemsForDisplay,
   sortPlaylists,
 } from '@/components/admin/playlists/playlist-helpers';
-import type { DropPlacement, EditablePlaylist, PlaylistItemRecord, PlaylistsManagerProps, PlaylistSummary } from '@/components/admin/playlists/playlist-types';
+import type { EditablePlaylist, PlaylistItemRecord, PlaylistsManagerProps, PlaylistSummary } from '@/components/admin/playlists/playlist-types';
 
 export function PlaylistsManager({
   initialPlaylists,
@@ -42,10 +36,6 @@ export function PlaylistsManager({
   const [selectedId, setSelectedId] = useState<string | null>(initialPlaylistId);
   const [items, setItems] = useState<PlaylistItemRecord[]>(() => sortItemsForDisplay(initialItems));
   const [isItemsDirty, setItemsDirty] = useState(false);
-  const [draggingId, setDraggingId] = useState<string | null>(null);
-  const [dropTargetId, setDropTargetId] = useState<string | null>(null);
-  const [dropAtEnd, setDropAtEnd] = useState(false);
-  const [dropPlacement, setDropPlacement] = useState<DropPlacement>('before');
   const [isPending, startTransition] = useTransition();
   const [feedback, setFeedback] = useState<string | null>(null);
   const [error, setError] = useState<string | null>(null);
@@ -56,20 +46,28 @@ export function PlaylistsManager({
   const [createName, setCreateName] = useState('');
   const [createSlug, setCreateSlug] = useState('');
   const [createDescription, setCreateDescription] = useState('');
-  const dragGhostRef = useRef<HTMLElement | null>(null);
-  const draggingIdRef = useRef<string | null>(null);
+  const {
+    clearDragState,
+    draggingId,
+    dropAtEnd,
+    dropPlacement,
+    dropTargetId,
+    handleCardDragOver,
+    handleDragEnd,
+    handleDragOver,
+    handleDragStart,
+    handleDropAtEnd,
+    handleDropOnCard,
+    handleDropOnPlaceholder,
+  } = usePlaylistDragReorder({ setItems, setItemsDirty });
 
   useEffect(() => {
     setPlaylists(sortPlaylists(initialPlaylists));
     setSelectedId(initialPlaylistId);
     setItems(sortItemsForDisplay(initialItems));
     setItemsDirty(false);
-    setDraggingId(null);
-    setDropTargetId(null);
-    setDropAtEnd(false);
-    setDropPlacement('before');
-    draggingIdRef.current = null;
-  }, [initialItems, initialPlaylistId, initialPlaylists]);
+    clearDragState();
+  }, [clearDragState, initialItems, initialPlaylistId, initialPlaylists]);
 
   const selectedPlaylist = useMemo(
     () => playlists.find((playlist) => playlist.id === selectedId) ?? null,
@@ -93,21 +91,6 @@ export function PlaylistsManager({
   );
 
   const familyHelpers = useMemo(() => buildFamilyHelpers(playlists), [playlists]);
-
-  const clearDragState = useCallback(() => {
-    draggingIdRef.current = null;
-    setDraggingId(null);
-    setDropTargetId(null);
-    setDropAtEnd(false);
-    setDropPlacement('before');
-  }, []);
-
-  const clearDragGhost = useCallback(() => {
-    if (dragGhostRef.current) {
-      dragGhostRef.current.remove();
-      dragGhostRef.current = null;
-    }
-  }, []);
 
   const syncPlaylistDetail = useCallback(
     (playlistId: string, nextPlaylist: PlaylistSummary | undefined, nextItems: PlaylistItemRecord[], basePlaylists?: EditablePlaylist[]) => {
@@ -432,115 +415,6 @@ export function PlaylistsManager({
     });
   }, [refreshPlaylistsState, selectedId]);
 
-  const commitDragMove = useCallback((fromId: string, toId: string | null, placement: DropPlacement = 'before') => {
-    setItems((current) => {
-      const fromIndex = current.findIndex((item) => item.videoId === fromId);
-      if (fromIndex === -1) return current;
-      const next = [...current];
-      const [moved] = next.splice(fromIndex, 1);
-      const insertIndex = toId ? next.findIndex((item) => item.videoId === toId) : next.length;
-      const targetIndex = insertIndex === -1 ? next.length : placement === 'after' ? insertIndex + 1 : insertIndex;
-      next.splice(targetIndex, 0, moved);
-      return next.map((item, orderIndex) => ({ ...item, orderIndex }));
-    });
-    setItemsDirty(true);
-    clearDragState();
-  }, [clearDragState]);
-
-  const handleDragStart = useCallback((event: DragEvent<HTMLElement>, videoId: string) => {
-    clearDragGhost();
-    event.dataTransfer.effectAllowed = 'move';
-    event.dataTransfer.setData('text/plain', videoId);
-    const source = event.currentTarget;
-    const ghost = source.cloneNode(true) as HTMLElement;
-    ghost.style.position = 'fixed';
-    ghost.style.top = '-10000px';
-    ghost.style.left = '-10000px';
-    ghost.style.width = `${source.clientWidth}px`;
-    ghost.style.pointerEvents = 'none';
-    ghost.style.opacity = '0.92';
-    ghost.style.transform = 'scale(0.98)';
-    ghost.style.boxShadow = '0 24px 60px rgba(0, 0, 0, 0.28)';
-    ghost.style.borderRadius = '20px';
-    ghost.style.overflow = 'hidden';
-    ghost.style.zIndex = '9999';
-    document.body.appendChild(ghost);
-    dragGhostRef.current = ghost;
-    event.dataTransfer.setDragImage(ghost, source.clientWidth / 2, 32);
-    draggingIdRef.current = videoId;
-    setDraggingId(videoId);
-    setDropTargetId(null);
-    setDropAtEnd(false);
-    setDropPlacement('before');
-  }, [clearDragGhost]);
-
-  const handleDragEnd = useCallback(() => {
-    clearDragGhost();
-    clearDragState();
-  }, [clearDragGhost, clearDragState]);
-
-  const handleDragOver = useCallback((event: DragEvent<HTMLElement>) => {
-    event.preventDefault();
-    event.dataTransfer.dropEffect = 'move';
-    if (event.target === event.currentTarget) {
-      setDropTargetId(null);
-      setDropAtEnd(true);
-      setDropPlacement('after');
-    }
-  }, []);
-
-  const handleDropOnCard = useCallback(
-    (event: DragEvent<HTMLElement>, targetVideoId: string) => {
-      event.preventDefault();
-      const activeDraggingId = draggingIdRef.current;
-      if (!activeDraggingId || activeDraggingId === targetVideoId) return;
-      const rect = event.currentTarget.getBoundingClientRect();
-      const nextPlacement: DropPlacement = event.clientY >= rect.top + rect.height / 2 ? 'after' : 'before';
-      commitDragMove(activeDraggingId, targetVideoId, nextPlacement);
-    },
-    [commitDragMove]
-  );
-
-  const handleDropAtEnd = useCallback(
-    (event: DragEvent<HTMLElement>) => {
-      if (event.target !== event.currentTarget) return;
-      event.preventDefault();
-      const activeDraggingId = draggingIdRef.current;
-      if (!activeDraggingId) return;
-      commitDragMove(activeDraggingId, null, 'after');
-    },
-    [commitDragMove]
-  );
-
-  const renderDropPlaceholder = useCallback(
-    (key: string, targetVideoId: string | null, placement: DropPlacement) => (
-      <div
-        key={key}
-        onDragOver={(event) => {
-          event.preventDefault();
-          event.stopPropagation();
-          event.dataTransfer.dropEffect = 'move';
-        }}
-        onDrop={(event) => {
-          event.preventDefault();
-          event.stopPropagation();
-          const activeDraggingId = draggingIdRef.current;
-          if (!activeDraggingId) return;
-          commitDragMove(activeDraggingId, targetVideoId, placement);
-        }}
-        className="rounded-card border-2 border-dashed border-brand/55 bg-brand/5 p-3 shadow-[inset_0_0_0_1px_rgba(0,0,0,0.02)]"
-      >
-        <div className="flex aspect-video items-center justify-center rounded-card border border-white/30 bg-gradient-to-br from-brand/10 to-brand/5 text-center">
-          <div className="space-y-1">
-            <p className="text-[11px] font-semibold uppercase tracking-[0.18em] text-brand/80">Drop here</p>
-            <p className="text-xs text-text-muted">New gallery position</p>
-          </div>
-        </div>
-      </div>
-    ),
-    [commitDragMove]
-  );
-
   const handleSaveItems = useCallback(() => {
     if (!selectedId) return;
     startTransition(async () => {
@@ -572,204 +446,47 @@ export function PlaylistsManager({
 
   return (
     <div className={clsx('space-y-8', isItemsDirty && 'pb-28', className)}>
-      <header className="flex flex-col gap-4 lg:flex-row lg:items-end lg:justify-between">
-        {!embedded ? (
-          <div className="space-y-2">
-            <p className="text-xs font-semibold uppercase tracking-micro text-text-muted">Admin</p>
-            <h1 className="text-3xl font-semibold text-text-primary">Collections curation</h1>
-            <p className="max-w-3xl text-sm text-text-secondary">
-              Keep the runtime surfaces stable while preparing family playlists. The main examples hub stays on <code>examples</code>,
-              guest starter stays on <code>welcome</code>, and family pages can be seeded from the current live order before you
-              start reordering them.
-            </p>
-          </div>
-        ) : null}
-        <div className="flex flex-wrap gap-2">
-          <Button type="button" variant="outline" size="sm" onClick={() => setShowDraftCollections((current) => !current)}>
-            {showDraftCollections ? 'Hide draft / empty' : `Show draft / empty (${groupedPlaylists.draft.length})`}
-          </Button>
-          <Button type="button" variant="outline" size="sm" onClick={() => handleCreateMissingFamilyPlaylists()}>
-            {missingFamilyCount > 0 ? `Create missing family playlists (${missingFamilyCount})` : 'Sync family playlists'}
-          </Button>
-          <Button type="button" variant="outline" size="sm" onClick={handleSeedAllFamilyPlaylists}>
-            Seed all family playlists
-          </Button>
-          <Button type="button" size="sm" onClick={() => setShowCreateForm((current) => !current)}>
-            {showCreateForm ? 'Close new collection' : '+ New collection'}
-          </Button>
-        </div>
-      </header>
-
-      {showCreateForm ? (
-        <PlaylistCreateForm
-          createDescription={createDescription}
-          createName={createName}
-          createSlug={createSlug}
-          isPending={isPending}
-          onCancel={() => setShowCreateForm(false)}
-          onDescriptionChange={setCreateDescription}
-          onNameChange={(nextName) => {
-            setCreateName(nextName);
-            if (!createSlug.trim()) {
-              setCreateSlug(slugify(nextName));
-            }
-          }}
-          onSlugChange={(value) => setCreateSlug(slugify(value))}
-          onSubmit={handleCreatePlaylist}
-        />
-      ) : null}
+      <PlaylistsManagerToolbar
+        createDescription={createDescription} createName={createName} createSlug={createSlug}
+        draftCount={groupedPlaylists.draft.length} embedded={embedded} isPending={isPending}
+        missingFamilyCount={missingFamilyCount} onCreateDescriptionChange={setCreateDescription}
+        onCreateMissingFamilyPlaylists={() => handleCreateMissingFamilyPlaylists()}
+        onCreateNameChange={setCreateName} onCreateSlugChange={setCreateSlug} onCreateSubmit={handleCreatePlaylist}
+        onSeedAllFamilyPlaylists={handleSeedAllFamilyPlaylists} showCreateForm={showCreateForm}
+        onToggleCreateForm={() => setShowCreateForm((current) => !current)}
+        onToggleDraftCollections={() => setShowDraftCollections((current) => !current)}
+        showDraftCollections={showDraftCollections}
+      />
 
       <PlaylistFeedbackBanners error={error} feedback={feedback} />
 
       <div className="grid gap-6 xl:grid-cols-[360px_minmax(0,1fr)]">
         <PlaylistsSidebar
-          emptyFamilyCount={emptyFamilyCount}
-          familyHelpers={familyHelpers}
-          groupedPlaylists={groupedPlaylists}
+          emptyFamilyCount={emptyFamilyCount} familyHelpers={familyHelpers} groupedPlaylists={groupedPlaylists}
           onCreateMissingFamilyPlaylists={handleCreateMissingFamilyPlaylists}
-          onSeedFamilyPlaylist={handleSeedFamilyPlaylist}
-          onSelectPlaylist={handleSelectPlaylist}
+          onSeedFamilyPlaylist={handleSeedFamilyPlaylist} onSelectPlaylist={handleSelectPlaylist}
           onToggleFamilyCollections={() => setShowFamilyCollections((current) => !current)}
           onToggleModelCollections={() => setShowModelCollections((current) => !current)}
-          pending={isPending}
-          selectedId={selectedId}
-          showDraftCollections={showDraftCollections}
-          showFamilyCollections={showFamilyCollections}
-          showModelCollections={showModelCollections}
+          pending={isPending} selectedId={selectedId} showDraftCollections={showDraftCollections}
+          showFamilyCollections={showFamilyCollections} showModelCollections={showModelCollections}
         />
 
         <section className="space-y-6">
-          {selectedPlaylist ? (
-            <>
-              <PlaylistDetailsPanel
-                isPending={isPending}
-                onDeletePlaylist={handleDeletePlaylist}
-                onFieldChange={handleFieldChange}
-                onSavePlaylist={handleSavePlaylist}
-                onSeedFamilyPlaylist={handleSeedFamilyPlaylist}
-                playlist={selectedPlaylist}
-              />
-
-              <section className="rounded-card border border-border bg-surface p-6 shadow-card">
-                <div className="flex flex-col gap-4 lg:flex-row lg:items-end lg:justify-between">
-                  <div>
-                    <h2 className="text-lg font-semibold text-text-primary">Collection items</h2>
-                    <p className="text-sm text-text-secondary">
-                      Drag clips to change the visible order, then save the gallery. Family playlists define the editorial top of the page.
-                    </p>
-                  </div>
-                  <Button
-                    type="button"
-                    size="sm"
-                    onClick={handleSaveItems}
-                    disabled={!isItemsDirty || isPending}
-                    className={clsx(isItemsDirty && 'shadow-card ring-2 ring-brand/20')}
-                  >
-                    Save order
-                  </Button>
-                </div>
-
-                {items.length ? (
-                  <div
-                    className="mt-5 grid gap-4 md:grid-cols-2 xl:grid-cols-3 2xl:grid-cols-4"
-                    onDragOver={handleDragOver}
-                    onDrop={handleDropAtEnd}
-                  >
-                    {items.map((item, index) => (
-                      <Fragment key={item.videoId}>
-                        {draggingId && dropTargetId === item.videoId && dropPlacement === 'before'
-                          ? renderDropPlaceholder(`placeholder-before-${item.videoId}`, item.videoId, 'before')
-                          : null}
-                        <article
-                          draggable
-                          onDragStart={(event) => handleDragStart(event, item.videoId)}
-                          onDragEnd={handleDragEnd}
-                          onDragOver={(event) => {
-                            handleDragOver(event);
-                            const activeDraggingId = draggingIdRef.current;
-                            if (!activeDraggingId || activeDraggingId === item.videoId) return;
-                            const rect = event.currentTarget.getBoundingClientRect();
-                            const nextPlacement: DropPlacement = event.clientY >= rect.top + rect.height / 2 ? 'after' : 'before';
-                            setDropTargetId(item.videoId);
-                            setDropAtEnd(false);
-                            setDropPlacement(nextPlacement);
-                          }}
-                          onDrop={(event) => handleDropOnCard(event, item.videoId)}
-                          className={clsx(
-                            'cursor-grab overflow-hidden rounded-card border border-hairline bg-bg shadow-card transition active:cursor-grabbing',
-                            draggingId === item.videoId && 'scale-[0.985] opacity-35'
-                          )}
-                        >
-                          <div className="relative aspect-video overflow-hidden bg-placeholder">
-                            <Image
-                              src={item.thumbUrl || getPlaceholderThumb(item.aspectRatio)}
-                              alt={`Thumbnail for ${item.videoId}`}
-                              fill
-                              className="object-cover"
-                              unoptimized
-                            />
-                            <div className="absolute inset-x-0 top-0 flex items-start justify-between gap-3 bg-gradient-to-b from-black/70 via-black/25 to-transparent p-3">
-                              <StatusPill tone="neutral">#{index + 1}</StatusPill>
-                              <Button
-                                type="button"
-                                size="sm"
-                                variant="outline"
-                                onClick={() => handleRemoveVideo(item.videoId)}
-                                disabled={isPending}
-                                className="min-h-[32px] rounded-full border-white/35 bg-black/35 px-2.5 text-white hover:border-white/60 hover:bg-black/50 hover:text-white"
-                                aria-label={`Remove ${item.engineLabel ?? item.videoId} from collection`}
-                              >
-                                <svg viewBox="0 0 20 20" aria-hidden="true" className="h-4 w-4 fill-current">
-                                  <path d="M7.5 2.5h5a1 1 0 0 1 1 1V5H17a.75.75 0 0 1 0 1.5h-.64l-.62 9.23A2 2 0 0 1 13.75 17.5h-7.5a2 2 0 0 1-1.99-1.77L3.64 6.5H3A.75.75 0 0 1 3 5h3.5V3.5a1 1 0 0 1 1-1Zm4.5 2.5V4h-4v1h4Zm-3 3.25a.75.75 0 0 0-1.5 0v5a.75.75 0 0 0 1.5 0v-5Zm3.5-.75a.75.75 0 0 1 .75.75v5a.75.75 0 0 1-1.5 0v-5a.75.75 0 0 1 .75-.75Z" />
-                                </svg>
-                              </Button>
-                            </div>
-                          </div>
-
-                          <div className="px-4 py-4">
-                            <p className="truncate text-sm font-semibold text-text-primary">{item.engineLabel ?? 'Unknown model'}</p>
-                          </div>
-                        </article>
-                        {draggingId && dropTargetId === item.videoId && dropPlacement === 'after'
-                          ? renderDropPlaceholder(`placeholder-after-${item.videoId}`, item.videoId, 'after')
-                          : null}
-                      </Fragment>
-                    ))}
-                    {draggingId && dropAtEnd ? renderDropPlaceholder('placeholder-end', null, 'after') : null}
-                  </div>
-                ) : (
-                  <div className="mt-5 rounded-card border border-dashed border-hairline bg-bg px-4 py-8 text-center text-sm text-text-secondary">
-                    This collection is empty. Use the family seed helpers or add clips from moderation, then reorder them here.
-                  </div>
-                )}
-              </section>
-            </>
-          ) : (
-            <div className="rounded-card border border-dashed border-hairline bg-surface p-10 text-center text-sm text-text-secondary">
-              Select a collection from the left rail to start curating.
-            </div>
-          )}
+          <PlaylistsManagerSelectionPanel
+            draggingId={draggingId} dropAtEnd={dropAtEnd} dropPlacement={dropPlacement} dropTargetId={dropTargetId}
+            isItemsDirty={isItemsDirty} isPending={isPending} items={items}
+            onCardDragOver={handleCardDragOver} onDeletePlaylist={handleDeletePlaylist} onDragEnd={handleDragEnd}
+            onDragOver={handleDragOver} onDragStart={handleDragStart} onDropAtEnd={handleDropAtEnd}
+            onDropOnCard={handleDropOnCard} onDropOnPlaceholder={handleDropOnPlaceholder}
+            onFieldChange={handleFieldChange} onRemoveVideo={handleRemoveVideo} onSaveItems={handleSaveItems}
+            onSavePlaylist={handleSavePlaylist} onSeedFamilyPlaylist={handleSeedFamilyPlaylist}
+            playlist={selectedPlaylist}
+          />
         </section>
       </div>
 
       {selectedPlaylist && isItemsDirty ? (
-        <div className="pointer-events-none fixed inset-x-0 bottom-5 z-40 flex justify-center px-4">
-          <div className="pointer-events-auto flex w-full max-w-xl items-center justify-between gap-4 rounded-full border border-brand/30 bg-surface/95 px-4 py-3 shadow-[0_18px_50px_rgba(0,0,0,0.18)] backdrop-blur">
-            <div className="min-w-0">
-              <p className="text-sm font-semibold text-text-primary">Order changed</p>
-              <p className="truncate text-xs text-text-secondary">{selectedPlaylist.name}</p>
-            </div>
-            <Button
-              type="button"
-              size="sm"
-              onClick={handleSaveItems}
-              disabled={isPending}
-              className="shrink-0 shadow-card ring-2 ring-brand/20"
-            >
-              Save order
-            </Button>
-          </div>
-        </div>
+        <PlaylistOrderDirtyBar isPending={isPending} playlistName={selectedPlaylist.name} onSaveItems={handleSaveItems} />
       ) : null}
     </div>
   );
